Factor out identifier existence assertion handling
